chore: Uppercase comments and extract validateChildKeys function
